dispatch stateful oneembedding ops
