+++ oneflow/core/functional/impl/nn_functor.cpp
@@ -3853,206 +3853,6 @@ class FusedCrossFeatureInteractionFunctor {
   std::shared_ptr<OpExpr> op_;
 };
 
-class OneEmbeddingIdShuffleFunctor {
- public:
-  OneEmbeddingIdShuffleFunctor() {
-    op_table_ids_has_in_out_ = CHECK_JUST(one::OpBuilder("id_shuffle")
-                                              .Input("ids")
-                                              .Input("table_ids")
-                                              .Output("num_unique_matrix")
-                                              .Output("inverse_unique_partition_indices")
-                                              .Output("cur_rank_num_unique")
-                                              .Output("cur_rank_unique_ids")
-                                              .Output("cur_rank_unique_table_ids")
-                                              .Output("cur_rank_inverse_indices")
-                                              .Build());
-    op_table_ids_no_in_has_out_ = CHECK_JUST(one::OpBuilder("id_shuffle")
-                                                 .Input("ids")
-                                                 .Output("num_unique_matrix")
-                                                 .Output("inverse_unique_partition_indices")
-                                                 .Output("cur_rank_num_unique")
-                                                 .Output("cur_rank_unique_ids")
-                                                 .Output("cur_rank_unique_table_ids")
-                                                 .Output("cur_rank_inverse_indices")
-                                                 .Build());
-  }
-
-  Maybe<TensorTuple> operator()(const std::shared_ptr<one::Tensor>& ids,
-                                const Optional<one::Tensor>& table_ids, const int32_t& num_tables,
-                                const std::string& embedding_name) const {
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP("num_tables", "embedding_name");
-    attrs.SetAllAttrs(num_tables, embedding_name);
-    if (table_ids) {
-      return OpInterpUtil::Dispatch<TensorTuple>(*op_table_ids_has_in_out_, {ids, JUST(table_ids)},
-                                                 attrs);
-    } else {
-      return OpInterpUtil::Dispatch<TensorTuple>(*op_table_ids_no_in_has_out_, {ids}, attrs);
-    }
-  }
-
- private:
-  std::shared_ptr<OpExpr> op_table_ids_has_in_out_;
-  std::shared_ptr<OpExpr> op_table_ids_no_in_has_out_;
-};
-
-class OneEmbeddingEmbeddingShuffleFunctor {
- public:
-  OneEmbeddingEmbeddingShuffleFunctor() {
-    op_ = CHECK_JUST(one::OpBuilder("embedding_shuffle")
-                         .Input("cur_rank_embeddings")
-                         .Input("num_unique_matrix")
-                         .Input("cur_rank_inverse_indices")
-                         .Input("inverse_unique_partition_indices")
-                         .Output("embeddings")
-                         .Build());
-  }
-
-  Maybe<Tensor> operator()(const std::shared_ptr<one::Tensor>& cur_rank_embeddings,
-                           const std::shared_ptr<one::Tensor>& num_unique_matrix,
-                           const std::shared_ptr<one::Tensor>& cur_rank_inverse_indices,
-                           const std::shared_ptr<one::Tensor>& inverse_unique_partition_indices,
-                           const std::string& embedding_name) const {
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP("embedding_size", "embedding_name");
-    const int64_t num_axes = cur_rank_embeddings->shape()->NumAxes();
-    attrs.SetAllAttrs(cur_rank_embeddings->shape()->At(num_axes - 1), embedding_name);
-    return OpInterpUtil::Dispatch<Tensor>(
-        *op_,
-        {cur_rank_embeddings, num_unique_matrix, cur_rank_inverse_indices,
-         inverse_unique_partition_indices},
-        attrs);
-  }
-
- private:
-  std::shared_ptr<OpExpr> op_;
-};
-
-class OneEmbeddingEmbeddingGradientShuffleFunctor {
- public:
-  OneEmbeddingEmbeddingGradientShuffleFunctor() {
-    op_ = CHECK_JUST(one::OpBuilder("embedding_gradient_shuffle")
-                         .Input("embedding_grad")
-                         .Input("num_unique_matrix")
-                         .Input("cur_rank_inverse_indices")
-                         .Input("inverse_unique_partition_indices")
-                         .Output("cur_rank_unique_embedding_grad")
-                         .Build());
-  }
-
-  Maybe<Tensor> operator()(const std::shared_ptr<one::Tensor>& embedding_grad,
-                           const std::shared_ptr<one::Tensor>& num_unique_matrix,
-                           const std::shared_ptr<one::Tensor>& cur_rank_inverse_indices,
-                           const std::shared_ptr<one::Tensor>& inverse_unique_partition_indices,
-                           const std::string& embedding_name) const {
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP("embedding_size", "embedding_name");
-    const int64_t num_axes = embedding_grad->shape()->NumAxes();
-    attrs.SetAllAttrs(embedding_grad->shape()->At(num_axes - 1), embedding_name);
-    return OpInterpUtil::Dispatch<Tensor>(
-        *op_,
-        {embedding_grad, num_unique_matrix, cur_rank_inverse_indices,
-         inverse_unique_partition_indices},
-        attrs);
-  }
-
- private:
-  std::shared_ptr<OpExpr> op_;
-};
-
-class OneEmbeddingLookupFunctor {
- public:
-  OneEmbeddingLookupFunctor() {
-    op_ = CHECK_JUST(one::OpBuilder("embedding_lookup")
-                         .Input("num_unique_ids")
-                         .Input("unique_ids")
-                         .Input("table_ids")
-                         .Output("unique_values")
-                         .Build());
-  }
-
-  Maybe<Tensor> operator()(const std::shared_ptr<one::Tensor>& num_unique_ids,
-                           const std::shared_ptr<one::Tensor>& unique_ids,
-                           const std::shared_ptr<one::Tensor>& table_ids,
-                           const Symbol<DType>& dtype, const Symbol<DType>& embedding_dtype,
-                           const int64_t line_size, const int64_t embedding_size,
-                           const std::string& embedding_name, const std::string& embedding_tables,
-                           const std::string& state_initializer, const int64_t seed) const {
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP("dtype", "embedding_dtype", "line_size",
-                                                 "embedding_size", "embedding_name",
-                                                 "embedding_tables", "state_initializer", "seed");
-    attrs.SetAllAttrs(dtype->data_type(), embedding_dtype->data_type(), line_size, embedding_size,
-                      embedding_name, embedding_tables, state_initializer, seed);
-    return OpInterpUtil::Dispatch<Tensor>(*op_, {num_unique_ids, unique_ids, table_ids}, attrs);
-  }
-
- private:
-  std::shared_ptr<OpExpr> op_;
-};
-
-class OneEmbeddingFusedLookupFunctor {
- public:
-  OneEmbeddingFusedLookupFunctor() {
-    op_has_table_ids_ = CHECK_JUST(one::OpBuilder("one_embedding_fused_lookup")
-                                       .Input("shadow")
-                                       .Input("ids")
-                                       .Input("table_ids")
-                                       .Output("embeddings")
-                                       .Build());
-    op_no_table_ids_ = CHECK_JUST(one::OpBuilder("one_embedding_fused_lookup")
-                                      .Input("shadow")
-                                      .Input("ids")
-                                      .Output("embeddings")
-                                      .Build());
-  }
-
-  Maybe<Tensor> operator()(const std::shared_ptr<one::Tensor>& shadow,
-                           const std::shared_ptr<one::Tensor>& ids,
-                           const Optional<one::Tensor>& table_ids, const Symbol<DType>& dtype,
-                           const std::string& embedding_name, const int64_t line_size,
-                           const int64_t embedding_size, const bool is_full_cache,
-                           const int32_t num_tables, const std::string& embedding_tables,
-                           const Optional<int64_t>& padding_idx, const int64_t seed) const {
-    int64_t padding_idx_val = -1;
-    bool has_padding_idx = false;
-    if (padding_idx.has_value()) {
-      padding_idx_val = JUST(padding_idx);
-      has_padding_idx = true;
-    }
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP(
-        "dtype", "embedding_name", "line_size", "embedding_size", "is_full_cache", "num_tables",
-        "embedding_tables", "seed", "padding_idx", "has_padding_idx");
-    attrs.SetAllAttrs(dtype->data_type(), embedding_name, line_size, embedding_size, is_full_cache,
-                      num_tables, embedding_tables, seed, padding_idx_val, has_padding_idx);
-    if (table_ids) {
-      const auto& table_ids_shape = *(JUST(table_ids)->shape());
-      const auto& ids_shape = *(ids->shape());
-      auto broadcast_table_ids = JUST(table_ids);
-      if (table_ids_shape != ids_shape) {
-        CHECK_LE_OR_RETURN(table_ids_shape.NumAxes(), ids_shape.NumAxes())
-            << "table_ids num_axes should be less equal to ids num_axes, but got table_ids "
-               "num_axes "
-            << table_ids_shape.NumAxes() << " and ids num_axes " << ids_shape.NumAxes();
-        const int64_t left_extend_dims = ids_shape.NumAxes() - table_ids_shape.NumAxes();
-        for (int64_t i = 0; i < table_ids_shape.NumAxes(); i++) {
-          CHECK_EQ_OR_RETURN(table_ids_shape.at(i), ids_shape.at(left_extend_dims + i))
-              << "when table_ids's shape not equals ids shape, table_ids must be able to be "
-                 "broadcast to ids_shape "
-                 "but got table_ids_shape: "
-              << table_ids_shape.DebugStr() << ", ids_shape: " << ids_shape.DebugStr();
-        }
-        broadcast_table_ids =
-            JUST(functional::BroadcastLike(JUST(table_ids), ids, std::vector<int32_t>{}));
-      }
-      return OpInterpUtil::Dispatch<Tensor>(*op_has_table_ids_, {shadow, ids, broadcast_table_ids},
-                                            attrs);
-    } else {
-      return OpInterpUtil::Dispatch<Tensor>(*op_no_table_ids_, {shadow, ids}, attrs);
-    }
-  }
-
- private:
-  std::shared_ptr<OpExpr> op_has_table_ids_;
-  std::shared_ptr<OpExpr> op_no_table_ids_;
-};
-
 class OneEmbeddingFusedLookupGradFunctor {
  public:
   OneEmbeddingFusedLookupGradFunctor() {
@@ -4074,381 +3874,6 @@ class OneEmbeddingFusedLookupGradFunctor {
 
  private:
   std::shared_ptr<OpExpr> op_;
-};
-
-class OneEmbeddingEmbeddingPutFunctor {
- public:
-  OneEmbeddingEmbeddingPutFunctor() {
-    op_ = CHECK_JUST(one::OpBuilder("embedding_put")
-                         .Input("num_unique_ids")
-                         .Input("unique_ids")
-                         .Input("unique_embeddings")
-                         .Build());
-  }
-
-  Maybe<void> operator()(const std::shared_ptr<one::Tensor>& num_unique_ids,
-                         const std::shared_ptr<one::Tensor>& unique_ids,
-                         const std::shared_ptr<one::Tensor>& unique_embeddings,
-                         const std::string& embedding_name, const int64_t line_size) const {
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP("embedding_name", "line_size");
-    attrs.SetAllAttrs(embedding_name, line_size);
-    JUST(OpInterpUtil::Dispatch<TensorTuple>(*op_, {num_unique_ids, unique_ids, unique_embeddings},
-                                             attrs));
-    return Maybe<void>::Ok();
-  }
-
- private:
-  std::shared_ptr<OpExpr> op_;
-};
-
-class OneEmbeddingUniqueKeyValuePairFunctor {
- public:
-  OneEmbeddingUniqueKeyValuePairFunctor() {
-    op_has_input_value_ = CHECK_JUST(one::OpBuilder("unique_key_value_pair")
-                                         .Input("keys")
-                                         .Input("values")
-                                         .Output("num_unique")
-                                         .Output("unique_keys")
-                                         .Output("unique_values")
-                                         .Output("inverse_indices")
-                                         .Build());
-    op_no_input_value_ = CHECK_JUST(one::OpBuilder("unique_key_value_pair")
-                                        .Input("keys")
-                                        .Output("num_unique")
-                                        .Output("unique_keys")
-                                        .Output("unique_values")
-                                        .Output("inverse_indices")
-                                        .Build());
-  }
-
-  Maybe<TensorTuple> operator()(const std::shared_ptr<one::Tensor>& keys,
-                                const Optional<one::Tensor>& values, const int32_t num_tables,
-                                const std::string& embedding_name) const {
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP("num_tables", "embedding_name");
-    attrs.SetAllAttrs(num_tables, embedding_name);
-    if (values) {
-      return OpInterpUtil::Dispatch<TensorTuple>(*op_has_input_value_, {keys, JUST(values)}, attrs);
-    } else {
-      return OpInterpUtil::Dispatch<TensorTuple>(*op_no_input_value_, {keys}, attrs);
-    }
-  }
-
- private:
-  std::shared_ptr<OpExpr> op_has_input_value_;
-  std::shared_ptr<OpExpr> op_no_input_value_;
-};
-
-class OneEmbeddingSgdUpdateFunctor {
- public:
-  OneEmbeddingSgdUpdateFunctor() {
-    // This functor is only used in one_embedding eager mode with lr passed by attr and no optional
-    // input, we also define functor with all optional input just for unittest. when the optional
-    // input learning_rate tensor has passed in, we think all optional input are not None and check
-    // them.
-    sgd_no_optional_input_op_ = CHECK_JUST(one::OpBuilder("one_embedding_sgd_update")
-                                               .Input("num_unique_ids")
-                                               .Input("unique_embeddings")
-                                               .Input("embedding_grad")
-                                               .Output("updated_unique_embeddings")
-                                               .Build());
-    momentum_no_optional_input_op_ = CHECK_JUST(one::OpBuilder("one_embedding_momentum_update")
-                                                    .Input("num_unique_ids")
-                                                    .Input("unique_embeddings")
-                                                    .Input("embedding_grad")
-                                                    .Output("updated_unique_embeddings")
-                                                    .Build());
-    // This functor is just for unittest
-    sgd_op_ = CHECK_JUST(one::OpBuilder("one_embedding_sgd_update")
-                             .Input("num_unique_ids")
-                             .Input("unique_embeddings")
-                             .Input("embedding_grad")
-                             .Input("learning_rate")
-                             .Input("down_scale_by_tensor")
-                             .Input("skip_if")
-                             .Output("updated_unique_embeddings")
-                             .Build());
-    momentum_op_ = CHECK_JUST(one::OpBuilder("one_embedding_momentum_update")
-                                  .Input("num_unique_ids")
-                                  .Input("unique_embeddings")
-                                  .Input("embedding_grad")
-                                  .Input("learning_rate")
-                                  .Input("down_scale_by_tensor")
-                                  .Input("skip_if")
-                                  .Output("updated_unique_embeddings")
-                                  .Build());
-  }
-
-  Maybe<Tensor> operator()(const std::shared_ptr<one::Tensor>& num_unique_ids,
-                           const std::shared_ptr<one::Tensor>& unique_embeddings,
-                           const std::shared_ptr<one::Tensor>& embedding_grad,
-                           const Optional<one::Tensor>& learning_rate,
-                           const Optional<one::Tensor>& down_scale_by_tensor,
-                           const Optional<one::Tensor>& skip_if, const float learning_rate_val,
-                           const double scale, const float weight_decay, const float momentum,
-                           const int64_t line_size, const int64_t embedding_size,
-                           const std::string& embedding_name) const {
-    auto& attrs =
-        THREAD_CACHED_MUTABLE_ATTR_MAP("learning_rate_val", "scale", "weight_decay", "line_size",
-                                       "embedding_size", "embedding_name", "beta");
-    if (momentum == 0) {
-      attrs.SetAllAttrs(learning_rate_val, scale, weight_decay, line_size, embedding_size,
-                        embedding_name, NullOpt);
-
-      if (learning_rate) {
-        CHECK(down_scale_by_tensor);
-        CHECK(skip_if);
-        return OpInterpUtil::Dispatch<Tensor>(
-            *sgd_op_,
-            {num_unique_ids, unique_embeddings, embedding_grad, JUST(learning_rate),
-             JUST(down_scale_by_tensor), JUST(skip_if)},
-            attrs);
-      } else {
-        CHECK(!down_scale_by_tensor);
-        CHECK(!skip_if);
-        return OpInterpUtil::Dispatch<Tensor>(
-            *sgd_no_optional_input_op_, {num_unique_ids, unique_embeddings, embedding_grad}, attrs);
-      }
-    } else {
-      attrs.SetAllAttrs(learning_rate_val, scale, weight_decay, line_size, embedding_size,
-                        embedding_name, momentum);
-      if (learning_rate) {
-        CHECK(down_scale_by_tensor);
-        CHECK(skip_if);
-        return OpInterpUtil::Dispatch<Tensor>(
-            *momentum_op_,
-            {num_unique_ids, unique_embeddings, embedding_grad, JUST(learning_rate),
-             JUST(down_scale_by_tensor), JUST(skip_if)},
-            attrs);
-      } else {
-        CHECK(!down_scale_by_tensor);
-        CHECK(!skip_if);
-        return OpInterpUtil::Dispatch<Tensor>(*momentum_no_optional_input_op_,
-                                              {num_unique_ids, unique_embeddings, embedding_grad},
-                                              attrs);
-      }
-    }
-  }
-
- private:
-  std::shared_ptr<OpExpr> sgd_no_optional_input_op_;
-  std::shared_ptr<OpExpr> sgd_op_;
-  std::shared_ptr<OpExpr> momentum_no_optional_input_op_;
-  std::shared_ptr<OpExpr> momentum_op_;
-};
-
-class OneEmbeddingAdamUpdateFunctor {
- public:
-  OneEmbeddingAdamUpdateFunctor() {
-    // This functor is only used in one_embedding eager mode with lr passed by attr and no optional
-    // input, we also define functor with all optional input just for unittest. when the optional
-    // input learning_rate tensor has passed in, we think all optional input are not None and check
-    // them.
-    no_optional_input_op_ = CHECK_JUST(one::OpBuilder("one_embedding_adam_update")
-                                           .Input("num_unique_ids")
-                                           .Input("unique_embeddings")
-                                           .Input("embedding_grad")
-                                           .Output("updated_unique_embeddings")
-                                           .Build());
-    // This functor is just for unittest
-    no_bias_correction_op_ = CHECK_JUST(one::OpBuilder("one_embedding_adam_update")
-                                            .Input("num_unique_ids")
-                                            .Input("unique_embeddings")
-                                            .Input("embedding_grad")
-                                            .Input("learning_rate")
-                                            .Input("down_scale_by_tensor")
-                                            .Input("skip_if")
-                                            .Output("updated_unique_embeddings")
-                                            .Build());
-    do_bias_correction_op_ = CHECK_JUST(one::OpBuilder("one_embedding_adam_update")
-                                            .Input("num_unique_ids")
-                                            .Input("unique_embeddings")
-                                            .Input("embedding_grad")
-                                            .Input("learning_rate")
-                                            .Input("down_scale_by_tensor")
-                                            .Input("skip_if")
-                                            .Input("bias_correction1")
-                                            .Input("bias_correction2")
-                                            .Output("updated_unique_embeddings")
-                                            .Build());
-  }
-
-  Maybe<Tensor> operator()(
-      const std::shared_ptr<one::Tensor>& num_unique_ids,
-      const std::shared_ptr<one::Tensor>& unique_embeddings,
-      const std::shared_ptr<one::Tensor>& embedding_grad,
-      const Optional<one::Tensor>& learning_rate, const Optional<one::Tensor>& down_scale_by_tensor,
-      const Optional<one::Tensor>& skip_if, const Optional<one::Tensor>& bias_correction1,
-      const Optional<one::Tensor>& bias_correction2, const float learning_rate_val,
-      const double scale, const float weight_decay, const float beta1, const float beta2,
-      const float& bias_correction1_val, const float& bias_correction2_val, const float epsilon,
-      const bool do_bias_correction, const int64_t line_size, const int64_t embedding_size,
-      const std::string& embedding_name) const {
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP(
-        "learning_rate_val", "scale", "weight_decay", "beta1", "beta2", "epsilon",
-        "bias_correction1_val", "bias_correction2_val", "do_bias_correction", "line_size",
-        "embedding_size", "embedding_name");
-    attrs.SetAllAttrs(learning_rate_val, scale, weight_decay, beta1, beta2, epsilon,
-                      bias_correction1_val, bias_correction2_val, do_bias_correction, line_size,
-                      embedding_size, embedding_name);
-    if (learning_rate) {
-      CHECK(down_scale_by_tensor);
-      CHECK(skip_if);
-      if (do_bias_correction) {
-        CHECK(bias_correction1);
-        CHECK(bias_correction2);
-        return OpInterpUtil::Dispatch<Tensor>(
-            *do_bias_correction_op_,
-            {num_unique_ids, unique_embeddings, embedding_grad, JUST(learning_rate),
-             JUST(down_scale_by_tensor), JUST(skip_if), JUST(bias_correction1),
-             JUST(bias_correction2)},
-            attrs);
-      } else {
-        return OpInterpUtil::Dispatch<Tensor>(
-            *no_bias_correction_op_,
-            {num_unique_ids, unique_embeddings, embedding_grad, JUST(learning_rate),
-             JUST(down_scale_by_tensor), JUST(skip_if)},
-            attrs);
-      }
-    } else {
-      CHECK(!down_scale_by_tensor);
-      CHECK(!skip_if);
-      CHECK(!bias_correction1);
-      CHECK(!bias_correction2);
-      return OpInterpUtil::Dispatch<Tensor>(
-          *no_optional_input_op_, {num_unique_ids, unique_embeddings, embedding_grad}, attrs);
-    }
-  }
-
- private:
-  std::shared_ptr<OpExpr> no_bias_correction_op_;
-  std::shared_ptr<OpExpr> do_bias_correction_op_;
-  std::shared_ptr<OpExpr> no_optional_input_op_;
-};
-
-class OneEmbeddingAdagradUpdateFunctor {
- public:
-  OneEmbeddingAdagradUpdateFunctor() {
-    // This functor is only used in one_embedding eager mode with lr passed by attr and no optional
-    // input, we also define functor with all optional input just for unittest. when the optional
-    // input learning_rate tensor has passed in, we think all optional input are not None and check
-    // them.
-    op_no_optional_input_ = CHECK_JUST(one::OpBuilder("one_embedding_adagrad_update")
-                                           .Input("num_unique_ids")
-                                           .Input("unique_embeddings")
-                                           .Input("embedding_grad")
-                                           .Output("updated_unique_embeddings")
-                                           .Build());
-    // This functor is just for unittest
-    op_ = CHECK_JUST(one::OpBuilder("one_embedding_adagrad_update")
-                         .Input("num_unique_ids")
-                         .Input("unique_embeddings")
-                         .Input("embedding_grad")
-                         .Input("learning_rate")
-                         .Input("down_scale_by_tensor")
-                         .Input("skip_if")
-                         .Input("train_step")
-                         .Output("updated_unique_embeddings")
-                         .Build());
-  }
-
-  Maybe<Tensor> operator()(const std::shared_ptr<one::Tensor>& num_unique_ids,
-                           const std::shared_ptr<one::Tensor>& unique_embeddings,
-                           const std::shared_ptr<one::Tensor>& embedding_grad,
-                           const Optional<one::Tensor>& learning_rate,
-                           const Optional<one::Tensor>& down_scale_by_tensor,
-                           const Optional<one::Tensor>& skip_if,
-                           const Optional<one::Tensor>& train_step, const int64_t train_step_val,
-                           const float learning_rate_val, const double scale,
-                           const float weight_decay, const float lr_decay, const float epsilon,
-                           const int64_t line_size, const int64_t embedding_size,
-                           const std::string& embedding_name) const {
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP("train_step_val", "learning_rate_val", "scale",
-                                                 "weight_decay", "lr_decay", "epsilon", "line_size",
-                                                 "embedding_size", "embedding_name");
-    attrs.SetAllAttrs(train_step_val, learning_rate_val, scale, weight_decay, lr_decay, epsilon,
-                      line_size, embedding_size, embedding_name);
-    if (learning_rate) {
-      CHECK(down_scale_by_tensor);
-      CHECK(skip_if);
-      CHECK(train_step);
-      return OpInterpUtil::Dispatch<Tensor>(
-          *op_,
-          {num_unique_ids, unique_embeddings, embedding_grad, JUST(learning_rate),
-           JUST(down_scale_by_tensor), JUST(skip_if), JUST(train_step)},
-          attrs);
-    } else {
-      CHECK(!down_scale_by_tensor);
-      CHECK(!skip_if);
-      CHECK(!train_step);
-      return OpInterpUtil::Dispatch<Tensor>(
-          *op_no_optional_input_, {num_unique_ids, unique_embeddings, embedding_grad}, attrs);
-    }
-  }
-
- private:
-  std::shared_ptr<OpExpr> op_;
-  std::shared_ptr<OpExpr> op_no_optional_input_;
-};
-
-class OneEmbeddingFtrlUpdateFunctor {
- public:
-  OneEmbeddingFtrlUpdateFunctor() {
-    // This functor is only used in one_embedding eager mode with lr passed by attr and no optional
-    // input, we also define functor with all optional input just for unittest. when the optional
-    // input learning_rate tensor has passed in, we think all optional input are not None and check
-    // them.
-    op_no_optional_input_ = CHECK_JUST(one::OpBuilder("one_embedding_ftrl_update")
-                                           .Input("num_unique_ids")
-                                           .Input("unique_embeddings")
-                                           .Input("embedding_grad")
-                                           .Output("updated_unique_embeddings")
-                                           .Build());
-    // This functor is just for unittest
-    op_ = CHECK_JUST(one::OpBuilder("one_embedding_ftrl_update")
-                         .Input("num_unique_ids")
-                         .Input("unique_embeddings")
-                         .Input("embedding_grad")
-                         .Input("learning_rate")
-                         .Input("down_scale_by_tensor")
-                         .Input("skip_if")
-                         .Output("updated_unique_embeddings")
-                         .Build());
-  }
-
-  Maybe<Tensor> operator()(const std::shared_ptr<one::Tensor>& num_unique_ids,
-                           const std::shared_ptr<one::Tensor>& unique_embeddings,
-                           const std::shared_ptr<one::Tensor>& embedding_grad,
-                           const Optional<one::Tensor>& learning_rate,
-                           const Optional<one::Tensor>& down_scale_by_tensor,
-                           const Optional<one::Tensor>& skip_if, const float learning_rate_val,
-                           const double scale, const float weight_decay, const float lr_power,
-                           const float lambda1, const float lambda2, const float beta,
-                           const int64_t line_size, const int64_t embedding_size,
-                           const std::string& embedding_name) const {
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP("learning_rate_val", "scale", "weight_decay",
-                                                 "lr_power", "lambda1", "lambda2", "beta",
-                                                 "line_size", "embedding_size", "embedding_name");
-    attrs.SetAllAttrs(learning_rate_val, scale, weight_decay, lr_power, lambda1, lambda2, beta,
-                      line_size, embedding_size, embedding_name);
-    if (learning_rate) {
-      CHECK(down_scale_by_tensor);
-      CHECK(skip_if);
-      return OpInterpUtil::Dispatch<Tensor>(
-          *op_,
-          {num_unique_ids, unique_embeddings, embedding_grad, JUST(learning_rate),
-           JUST(down_scale_by_tensor), JUST(skip_if)},
-          attrs);
-    } else {
-      CHECK(!down_scale_by_tensor);
-      CHECK(!skip_if);
-      return OpInterpUtil::Dispatch<Tensor>(
-          *op_no_optional_input_, {num_unique_ids, unique_embeddings, embedding_grad}, attrs);
-    }
-  }
-
- private:
-  std::shared_ptr<OpExpr> op_;
-  std::shared_ptr<OpExpr> op_no_optional_input_;
 };
 
 class DeformConv2dFunctor {
@@ -5214,19 +4639,7 @@ ONEFLOW_FUNCTION_LIBRARY(m) {
   m.add_functor<impl::RoiAlignGradFunctor>("RoiAlignGrad");
   m.add_functor<impl::FusedDotFeatureInteractionFunctor>("FusedDotFeatureInteraction");
   m.add_functor<impl::FusedCrossFeatureInteractionFunctor>("FusedCrossFeatureInteraction");
-  m.add_functor<impl::OneEmbeddingIdShuffleFunctor>("OneEmbeddingIdShuffle");
-  m.add_functor<impl::OneEmbeddingEmbeddingShuffleFunctor>("OneEmbeddingEmbeddingShuffle");
-  m.add_functor<impl::OneEmbeddingEmbeddingGradientShuffleFunctor>(
-      "OneEmbeddingEmbeddingGradientShuffle");
-  m.add_functor<impl::OneEmbeddingLookupFunctor>("OneEmbeddingLookup");
-  m.add_functor<impl::OneEmbeddingFusedLookupFunctor>("OneEmbeddingFusedLookup");
   m.add_functor<impl::OneEmbeddingFusedLookupGradFunctor>("OneEmbeddingFusedLookupGrad");
-  m.add_functor<impl::OneEmbeddingEmbeddingPutFunctor>("OneEmbeddingEmbeddingPut");
-  m.add_functor<impl::OneEmbeddingUniqueKeyValuePairFunctor>("OneEmbeddingUniqueKeyValuePair");
-  m.add_functor<impl::OneEmbeddingSgdUpdateFunctor>("OneEmbeddingSgdUpdate");
-  m.add_functor<impl::OneEmbeddingAdamUpdateFunctor>("OneEmbeddingAdamUpdate");
-  m.add_functor<impl::OneEmbeddingAdagradUpdateFunctor>("OneEmbeddingAdagradUpdate");
-  m.add_functor<impl::OneEmbeddingFtrlUpdateFunctor>("OneEmbeddingFtrlUpdate");
   m.add_functor<impl::RocAucScoreFunctor>("RocAucScore");
   m.add_functor<impl::MultiTensorSgdUpdateFunctor>("MultiTensorSgdUpdate");
   m.add_functor<impl::MultiTensorMomentumUpdateFunctor>("MultiTensorMomentumUpdate");
